This maybe fail on Windows agent but 🤷‍♂

Why do we have to wait for incremental for plus 50 minutes?

![image](https://user-images.githubusercontent.com/1661688/66436385-36787980-ea27-11e9-8c4b-abad0be6fca5.png)

Perhaps using something like https://github.com/jenkinsci/http-request-plugin or some better alternative?

Just for reference the build took 5 minutes, however been waiting for deploy for 1 hour and 20 minutes.

```
23:07:44  Pull request #57 updated
21:07:44 [2019-10-08T21:07:44.178Z] Connecting to https://api.github.com using jenkinsadmin/****** (GitHub access token for jenkinsadmin)
23:07:44  Connecting to https://api.github.com to check permissions of obtain list of casz for jenkinsci/hashicorp-vault-plugin
23:07:45  Obtained Jenkinsfile from b8e767fa3e9ef07170697919036aa204bfb6f167+d06a570038e145d5bc335bfe594a997958430530 (bb531da180ff4ae68b091dddb0549cab459e1d1f)
23:07:45  Running in Durability level: PERFORMANCE_OPTIMIZED
23:07:45  Loading library pipeline-library@master
23:07:45  Attempting to resolve master from remote references...
23:07:45  Found match: refs/heads/master revision 478d9a971e23d8ec5ca57019cd3f478f42b6968e
23:07:45  No credentials specified
23:07:45  Fetching changes from the remote Git repository
23:07:45  Fetching without tags
23:07:45  Checking out Revision 478d9a971e23d8ec5ca57019cd3f478f42b6968e (master)
23:07:45  Commit message: "Merge pull request #112 from jenkins-infra/INFRA-2243-fix"
23:07:45  [Pipeline] Start of Pipeline
23:07:46  [Pipeline] properties
23:07:46  [Pipeline] parallel
23:07:46  [Pipeline] { (Branch: linux-8)
23:07:46  [Pipeline] { (Branch: linux-8-2.164.1)
23:07:46  [Pipeline] { (Branch: linux-11-2.164.1)
23:07:46  [Pipeline] node
23:07:46  [Pipeline] node
23:07:46  [Pipeline] node
23:07:46  Running on ubuntu-jenkinsinfra493cb0 in /home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57
23:07:46  [Pipeline] {
23:07:46  Running on ubuntu-jenkinsinfra921110 in /home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57
23:07:46  [Pipeline] {
23:07:46  [Pipeline] timeout
23:07:46  Timeout set to expire in 1 hr 0 min
23:07:46  [Pipeline] {
23:07:46  [Pipeline] timeout
23:07:46  Timeout set to expire in 1 hr 0 min
23:07:46  [Pipeline] {
23:07:46  [Pipeline] stage
23:07:46  [Pipeline] { (Checkout (linux-8))
23:07:46  [Pipeline] stage
23:07:46  [Pipeline] { (Checkout (linux-8-2.164.1))
23:07:46  Running on ubuntu-jenkinsinfra47c890 in /home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57
23:07:46  [Pipeline] checkout
23:07:46  [Pipeline] checkout
23:07:46  [Pipeline] {
23:07:46  [Pipeline] timeout
23:07:46  Timeout set to expire in 1 hr 0 min
23:07:46  [Pipeline] {
23:07:46  [Pipeline] stage
23:07:46  [Pipeline] { (Checkout (linux-11-2.164.1))
23:07:46  using credential github-access-token
23:07:46  [Pipeline] checkout
23:07:46  using credential github-access-token
23:07:46  using credential github-access-token
23:07:46  Cloning the remote Git repository
23:07:46  Cloning with configured refspecs honoured and without tags
23:07:46  Cloning the remote Git repository
23:07:46  Cloning with configured refspecs honoured and without tags
23:07:46  Cloning the remote Git repository
23:07:46  Cloning with configured refspecs honoured and without tags
23:07:46  Fetching without tags
23:07:46  Merging remotes/origin/master commit d06a570038e145d5bc335bfe594a997958430530 into PR head commit b8e767fa3e9ef07170697919036aa204bfb6f167
23:07:46  Fetching without tags
23:07:46  Merge succeeded, producing b8e767fa3e9ef07170697919036aa204bfb6f167
23:07:46  Checking out Revision b8e767fa3e9ef07170697919036aa204bfb6f167 (PR-57)
23:07:46  Merging remotes/origin/master commit d06a570038e145d5bc335bfe594a997958430530 into PR head commit b8e767fa3e9ef07170697919036aa204bfb6f167
23:07:46  Fetching without tags
23:07:46  Commit message: "fix assertion"
23:07:46  Merging remotes/origin/master commit d06a570038e145d5bc335bfe594a997958430530 into PR head commit b8e767fa3e9ef07170697919036aa204bfb6f167
23:07:46  Merge succeeded, producing b8e767fa3e9ef07170697919036aa204bfb6f167
23:07:46  Checking out Revision b8e767fa3e9ef07170697919036aa204bfb6f167 (PR-57)
23:07:47  Commit message: "fix assertion"
23:07:47  Merge succeeded, producing b8e767fa3e9ef07170697919036aa204bfb6f167
23:07:47  Checking out Revision b8e767fa3e9ef07170697919036aa204bfb6f167 (PR-57)
23:07:47  Commit message: "fix assertion"
23:07:46  remote: Enumerating objects
23:07:46  remote: Counting objects
23:07:46  remote: Compressing objects
23:07:46  Receiving objects
23:07:46  Resolving deltas
23:07:46  Updating references
23:07:46  remote: Enumerating objects
23:07:46  remote: Counting objects
23:07:46  remote: Compressing objects
23:07:46  Receiving objects
23:07:46  Resolving deltas
23:07:46  Updating references
23:07:46  remote: Enumerating objects
23:07:46  remote: Counting objects
23:07:46  remote: Compressing objects
23:07:46  Receiving objects
23:07:46  Resolving deltas
23:07:46  Updating references
23:07:47  [Pipeline] fileExists
23:07:47  [Pipeline] fileExists
23:07:47  [Pipeline] readFile
23:07:47  [Pipeline] }
23:07:47  [Pipeline] // stage
23:07:47  [Pipeline] stage
23:07:47  [Pipeline] { (Build (linux-8-2.164.1))
23:07:47  [Pipeline] pwd
23:07:47  [Pipeline] pwd
23:07:47  [Pipeline] fileExists
23:07:47  [Pipeline] libraryResource
23:07:47  [Pipeline] writeFile
23:07:47  [Pipeline] fileExists
23:07:47  [Pipeline] tool
23:07:47  [Pipeline] readFile
23:07:47  [Pipeline] }
23:07:47  [Pipeline] tool
23:07:47  [Pipeline] // stage
23:07:47  [Pipeline] stage
23:07:47  [Pipeline] { (Build (linux-8))
23:07:47  [Pipeline] withEnv
23:07:47  [Pipeline] {
23:07:47  [Pipeline] pwd
23:07:47  [Pipeline] pwd
23:07:47  [Pipeline] pwd
23:07:47  [Pipeline] isUnix
23:07:47  [Pipeline] sh
23:07:47  [Pipeline] libraryResource
23:07:47  [Pipeline] writeFile
23:07:47  [Pipeline] tool
23:07:47  [Pipeline] tool
23:07:47  [Pipeline] withEnv
23:07:47  [Pipeline] {
23:07:47  [Pipeline] isUnix
23:07:47  [Pipeline] sh
23:07:47  [Pipeline] fileExists
23:07:47  [Pipeline] fileExists
23:07:47  [Pipeline] readFile
23:07:47  [Pipeline] }
23:07:47  [Pipeline] // stage
23:07:47  [Pipeline] stage
23:07:47  [Pipeline] { (Build (linux-11-2.164.1))
23:07:47  [Pipeline] pwd
23:07:47  [Pipeline] pwd
23:07:47  [Pipeline] libraryResource
23:07:47  [Pipeline] writeFile
23:07:47  [Pipeline] tool
23:07:47  [Pipeline] tool
23:07:47  + mvn --batch-mode --show-version --errors -Dorg.slf4j.simpleLogger.log.org.apache.maven.cli.transfer.Slf4jMavenTransferListener=warn -s /home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57@tmp/settings-azure.xml --update-snapshots -Dmaven.repo.local=/home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57@tmp/m2repo -Dmaven.test.failure.ignore -Dset.changelist -Djenkins.version=2.164.1 -Daccess-modifier-checker.failOnError=false -Djava.level=8 -Dfindbugs.failOnError=false clean install
23:07:47  Apache Maven 3.5.4 (1edded0938998edf8bf061f1ceb3cfdeccf443fe; 2018-06-17T18:33:14Z)
23:07:47  Maven home: /home/jenkins/tools/hudson.tasks.Maven_MavenInstallation/mvn
23:07:47  Java version: 1.8.0_162, vendor: Oracle Corporation, runtime: /home/jenkins/tools/hudson.model.JDK/jdk8/jre
23:07:47  Default locale: en, platform encoding: UTF-8
23:07:47  OS name: "linux", version: "5.0.0-1018-azure", arch: "amd64", family: "unix"
23:07:47  + mvn --batch-mode --show-version --errors -Dorg.slf4j.simpleLogger.log.org.apache.maven.cli.transfer.Slf4jMavenTransferListener=warn -s /home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57@tmp/settings-azure.xml --update-snapshots -Dmaven.repo.local=/home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57@tmp/m2repo -Dmaven.test.failure.ignore -Dset.changelist help:evaluate -Dexpression=changelist -Doutput=/home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57@tmp/changelist -Dfindbugs.failOnError=false clean install
23:07:47  Apache Maven 3.5.4 (1edded0938998edf8bf061f1ceb3cfdeccf443fe; 2018-06-17T18:33:14Z)
23:07:47  Maven home: /home/jenkins/tools/hudson.tasks.Maven_MavenInstallation/mvn
23:07:47  Java version: 1.8.0_162, vendor: Oracle Corporation, runtime: /home/jenkins/tools/hudson.model.JDK/jdk8/jre
23:07:47  Default locale: en, platform encoding: UTF-8
23:07:47  OS name: "linux", version: "5.0.0-1018-azure", arch: "amd64", family: "unix"
23:07:47  Unpacking https://github.com/AdoptOpenJDK/openjdk11-binaries/releases/download/jdk-11.0.3%2B7/OpenJDK11U-jdk_x64_linux_hotspot_11.0.3_7.tar.gz to /home/jenkins/tools/hudson.model.JDK/jdk11 on ubuntu-jenkinsinfra47c890
23:07:49  [INFO] Error stacktraces are turned on.
23:07:50  [INFO] Setting: -Dchangelist=-rc263.b8e767fa3e9e -DscmTag=b8e767fa3e9ef07170697919036aa204bfb6f167
23:07:50  [INFO] Scanning for projects...
23:07:50  [INFO] Error stacktraces are turned on.
23:07:50  [INFO] Setting: -Dchangelist=-rc263.b8e767fa3e9e -DscmTag=b8e767fa3e9ef07170697919036aa204bfb6f167
23:07:50  [INFO] Scanning for projects...
23:07:51  [INFO] 
23:07:51  [INFO] --------< com.datapipe.jenkins.plugins:hashicorp-vault-plugin >---------
23:07:51  [INFO] Building HashiCorp Vault Plugin 2.5.1-rc263.b8e767fa3e9e
23:07:51  [INFO] --------------------------------[ hpi ]---------------------------------
23:07:52  [INFO] 
23:07:52  [INFO] --------< com.datapipe.jenkins.plugins:hashicorp-vault-plugin >---------
23:07:52  [INFO] Building HashiCorp Vault Plugin 2.5.1-rc263.b8e767fa3e9e
23:07:52  [INFO] --------------------------------[ hpi ]---------------------------------
23:07:52  [INFO] 
23:07:52  [INFO] --- maven-help-plugin:3.2.0:evaluate (default-cli) @ hashicorp-vault-plugin ---
23:07:53  [INFO] No artifact parameter specified, using 'com.datapipe.jenkins.plugins:hashicorp-vault-plugin:hpi:2.5.1-rc263.b8e767fa3e9e' as project.
23:07:53  [INFO] Result of evaluation written to: /home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57@tmp/changelist
23:07:53  [INFO] 
23:07:53  [INFO] --------< com.datapipe.jenkins.plugins:hashicorp-vault-plugin >---------
23:07:53  [INFO] Building HashiCorp Vault Plugin 2.5.1-rc263.b8e767fa3e9e
23:07:53  [INFO] --------------------------------[ hpi ]---------------------------------
23:07:53  [INFO] 
23:07:53  [INFO] --- maven-clean-plugin:3.1.0:clean (default-clean) @ hashicorp-vault-plugin ---
23:07:53  [INFO] 
23:07:53  [INFO] --- maven-hpi-plugin:3.6:validate (default-validate) @ hashicorp-vault-plugin ---
23:07:53  [INFO] 
23:07:53  [INFO] --- maven-enforcer-plugin:3.0.0-M2:display-info (display-info) @ hashicorp-vault-plugin ---
23:07:53  [INFO] Maven Version: 3.5.4
23:07:53  [INFO] JDK Version: 1.8.0_162 normalized as: 1.8.0-162
23:07:53  [INFO] OS Info: Arch: amd64 Family: unix Name: linux Version: 5.0.0-1018-azure
23:07:53  [INFO] 
23:07:53  [INFO] --- maven-enforcer-plugin:3.0.0-M2:enforce (display-info) @ hashicorp-vault-plugin ---
23:07:53  [INFO] 
23:07:53  [INFO] --- maven-clean-plugin:3.1.0:clean (default-clean) @ hashicorp-vault-plugin ---
23:07:53  [INFO] 
23:07:53  [INFO] --- maven-hpi-plugin:3.6:validate (default-validate) @ hashicorp-vault-plugin ---
23:07:53  [INFO] 
23:07:53  [INFO] --- maven-enforcer-plugin:3.0.0-M2:display-info (display-info) @ hashicorp-vault-plugin ---
23:07:53  [INFO] artifact junit:junit-dep: checking for updates from azure-incrementals
23:07:53  [INFO] Maven Version: 3.5.4
23:07:53  [INFO] JDK Version: 1.8.0_162 normalized as: 1.8.0-162
23:07:53  [INFO] OS Info: Arch: amd64 Family: unix Name: linux Version: 5.0.0-1018-azure
23:07:53  [INFO] 
23:07:53  [INFO] --- maven-enforcer-plugin:3.0.0-M2:enforce (display-info) @ hashicorp-vault-plugin ---
23:07:54  [INFO] artifact junit:junit-dep: checking for updates from azure
23:07:54  [INFO] artifact junit:junit-dep: checking for updates from central
23:07:54  [INFO] artifact junit:junit: checking for updates from azure-incrementals
23:07:54  [INFO] artifact junit:junit: checking for updates from azure
23:07:54  [INFO] artifact junit:junit-dep: checking for updates from azure-incrementals
23:07:54  [INFO] artifact junit:junit-dep: checking for updates from azure
23:07:54  [INFO] artifact junit:junit: checking for updates from central
23:07:54  [INFO] artifact junit:junit-dep: checking for updates from central
23:07:54  [INFO] artifact junit:junit: checking for updates from azure-incrementals
23:07:54  [INFO] artifact junit:junit: checking for updates from azure
23:07:54  [Pipeline] withEnv
23:07:54  [Pipeline] {
23:07:54  [Pipeline] isUnix
23:07:54  [Pipeline] sh
23:07:54  [INFO] artifact junit:junit: checking for updates from central
23:07:54  + mvn --batch-mode --show-version --errors -Dorg.slf4j.simpleLogger.log.org.apache.maven.cli.transfer.Slf4jMavenTransferListener=warn -s /home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57@tmp/settings-azure.xml --update-snapshots -Dmaven.repo.local=/home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57@tmp/m2repo -Dmaven.test.failure.ignore -Dset.changelist -Djenkins.version=2.164.1 -Daccess-modifier-checker.failOnError=false -Djava.level=8 -Dfindbugs.failOnError=false clean install
23:07:55  Apache Maven 3.5.4 (1edded0938998edf8bf061f1ceb3cfdeccf443fe; 2018-06-17T18:33:14Z)
23:07:55  Maven home: /home/jenkins/tools/hudson.tasks.Maven_MavenInstallation/mvn
23:07:55  Java version: 11.0.3, vendor: AdoptOpenJDK, runtime: /home/jenkins/tools/hudson.model.JDK/jdk11/jdk-11.0.3+7
23:07:55  Default locale: en, platform encoding: UTF-8
23:07:55  OS name: "linux", version: "5.0.0-1018-azure", arch: "amd64", family: "unix"
23:07:55  [INFO] Adding ignore: module-info
23:07:55  [INFO] artifact junit:junit-dep: checking for updates from azure-incrementals
23:07:55  [INFO] artifact junit:junit-dep: checking for updates from azure
23:07:56  [INFO] artifact junit:junit-dep: checking for updates from central
23:07:56  [INFO] artifact junit:junit: checking for updates from azure-incrementals
23:07:56  [INFO] artifact junit:junit: checking for updates from azure
23:07:56  [INFO] artifact junit:junit: checking for updates from central
23:07:56  [INFO] Adding ignore: module-info
23:07:56  [INFO] Ignoring requireUpperBoundDeps in net.java.dev.jna:jna
23:07:56  [INFO] 
23:07:56  [INFO] --- maven-localizer-plugin:1.26:generate (default) @ hashicorp-vault-plugin ---
23:07:56  [INFO] 
23:07:56  [INFO] --- maven-resources-plugin:3.1.0:resources (default-resources) @ hashicorp-vault-plugin ---
23:07:56  [INFO] Using 'UTF-8' encoding to copy filtered resources.
23:07:56  [INFO] Copying 27 resources
23:07:56  [INFO] 
23:07:56  [INFO] --- flatten-maven-plugin:1.1.0:flatten (flatten) @ hashicorp-vault-plugin ---
23:07:56  [INFO] Generating flattened POM of project com.datapipe.jenkins.plugins:hashicorp-vault-plugin:hpi:2.5.1-rc263.b8e767fa3e9e...
23:07:56  [INFO] 
23:07:56  [INFO] --- maven-compiler-plugin:3.8.1:compile (default-compile) @ hashicorp-vault-plugin ---
23:07:56  [INFO] Changes detected - recompiling the module!
23:07:56  [INFO] Compiling 29 source files to /home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57/target/classes
23:07:57  [INFO] artifact junit:junit-dep: checking for updates from azure-incrementals
23:07:57  [INFO] artifact junit:junit-dep: checking for updates from azure
23:07:57  [INFO] artifact junit:junit-dep: checking for updates from central
23:07:57  [INFO] artifact junit:junit: checking for updates from azure-incrementals
23:07:57  [INFO] artifact junit:junit: checking for updates from azure
23:07:57  [INFO] artifact junit:junit: checking for updates from central
23:07:57  [INFO] Ignoring requireUpperBoundDeps in net.java.dev.jna:jna
23:07:57  [INFO] 
23:07:57  [INFO] --- maven-localizer-plugin:1.26:generate (default) @ hashicorp-vault-plugin ---
23:07:57  [INFO] 
23:07:57  [INFO] --- maven-resources-plugin:3.1.0:resources (default-resources) @ hashicorp-vault-plugin ---
23:07:57  [INFO] Using 'UTF-8' encoding to copy filtered resources.
23:07:57  [INFO] Copying 27 resources
23:07:57  [INFO] 
23:07:57  [INFO] --- flatten-maven-plugin:1.1.0:flatten (flatten) @ hashicorp-vault-plugin ---
23:07:57  [INFO] Generating flattened POM of project com.datapipe.jenkins.plugins:hashicorp-vault-plugin:hpi:2.5.1-rc263.b8e767fa3e9e...
23:07:58  [INFO] 
23:07:58  [INFO] --- maven-compiler-plugin:3.8.1:compile (default-compile) @ hashicorp-vault-plugin ---
23:07:58  [INFO] Changes detected - recompiling the module!
23:07:58  [INFO] Compiling 29 source files to /home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57/target/classes
23:07:58  [INFO] /home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57/src/main/java/com/datapipe/jenkins/vault/credentials/VaultTokenFileCredential.java: Some input files use or override a deprecated API.
23:07:58  [INFO] /home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57/src/main/java/com/datapipe/jenkins/vault/credentials/VaultTokenFileCredential.java: Recompile with -Xlint:deprecation for details.
23:07:58  [INFO] 
23:07:58  [INFO] --- access-modifier-checker:1.16:enforce (default-enforce) @ hashicorp-vault-plugin ---
23:07:58  [INFO] 
23:07:58  [INFO] --- animal-sniffer-maven-plugin:1.17:check (check) @ hashicorp-vault-plugin ---
23:07:58  [INFO] Resolved signature org.codehaus.mojo.signature:java18 version as 1.0 from dependencyManagement
23:07:58  [INFO] Checking unresolved references to org.codehaus.mojo.signature:java18:1.0
23:07:59  [INFO] /home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57/src/main/java/com/datapipe/jenkins/vault/credentials/common/VaultUsernamePasswordCredentialImpl.java: /home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57/src/main/java/com/datapipe/jenkins/vault/credentials/common/VaultUsernamePasswordCredentialImpl.java uses or overrides a deprecated API.
23:07:59  [INFO] /home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57/src/main/java/com/datapipe/jenkins/vault/credentials/common/VaultUsernamePasswordCredentialImpl.java: Recompile with -Xlint:deprecation for details.
23:07:59  [INFO] 
23:07:59  [INFO] --- access-modifier-checker:1.16:enforce (default-enforce) @ hashicorp-vault-plugin ---
23:08:00  [INFO] 
23:08:00  [INFO] --- animal-sniffer-maven-plugin:1.17:check (check) @ hashicorp-vault-plugin ---
23:08:00  [INFO] Resolved signature org.codehaus.mojo.signature:java18 version as 1.0 from dependencyManagement
23:08:00  [INFO] Checking unresolved references to org.codehaus.mojo.signature:java18:1.0
23:08:00  [INFO] Error stacktraces are turned on.
23:08:00  [INFO] Setting: -Dchangelist=-rc263.b8e767fa3e9e -DscmTag=b8e767fa3e9ef07170697919036aa204bfb6f167
23:08:00  [INFO] Scanning for projects...
23:08:01  [INFO] 
23:08:01  [INFO] --- maven-hpi-plugin:3.6:insert-test (default-insert-test) @ hashicorp-vault-plugin ---
23:08:01  [INFO] 
23:08:01  [INFO] --- gmaven-plugin:1.5-jenkins-3:generateTestStubs (test-in-groovy) @ hashicorp-vault-plugin ---
23:08:01  [INFO] No sources found for Java stub generation
23:08:01  [INFO] 
23:08:01  [INFO] --- maven-antrun-plugin:1.3:run (createTempDir) @ hashicorp-vault-plugin ---
23:08:01  [INFO] Executing tasks
23:08:01      [mkdir] Created dir: /home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57/target/tmp
23:08:01  [INFO] Executed tasks
23:08:01  [INFO] 
23:08:01  [INFO] --- maven-resources-plugin:3.1.0:testResources (default-testResources) @ hashicorp-vault-plugin ---
23:08:01  [INFO] Using 'UTF-8' encoding to copy filtered resources.
23:08:01  [INFO] Copying 7 resources
23:08:01  [INFO] 
23:08:01  [INFO] --- maven-compiler-plugin:3.8.1:testCompile (default-testCompile) @ hashicorp-vault-plugin ---
23:08:01  [INFO] Changes detected - recompiling the module!
23:08:01  [INFO] Compiling 14 source files to /home/jenkins/workspace/ins_hashicorp-vault-plugin_PR-57/target/test-classes
…