@casz 2019-10-09
Most liekly related to, a build missing a timeout: https://github.com/jenkinsci/bom/pull/123