@jglick 2019-10-09
> Perhaps using something like https://github.com/jenkinsci/http-request-plugin

Yes that would be an option, assuming it can handle a slow response (`incrementals-publisher` sometimes runs for a few _minutes_ before giving a 200), if it were installed (there is no CasC for ci.jenkins.io so we cannot directly propose that).