@casz 2019-10-09
https://github.com/jenkinsci/http-request-plugin#features

It does support connection timeout