@casz 2019-10-09
a temporary option could be to surround the maybePublishIncrementals with a timeout that does not fail the build?