@jglick 2019-10-12
> It does support connection timeout

Well you can also wrap any step in a `timeout` block; I meant, will it _work_ even if the service takes several minutes to respond (after the connection has been established).

> surround the maybePublishIncrementals with a timeout that does not fail the build

Perhaps. The HTTP Client plugin sounds more attractive.

> a milestone

Not sure what you think that would accomplish.