@casz 2019-10-12
> Well you can also wrap any step in a `timeout` block; I meant, will it _work_ even if the service takes several minutes to respond (after the connection has been established).

The http request plugin supports http request timeout which could be set to 10 minutes of something reasonable.

> Perhaps. The HTTP Client plugin sounds more attractive.

With the HTTP client we would not need a node, we should be able to have the Jenkins Master send the request.

> Not sure what you think that would accomplish.

To avoid uploading old builds where a newer commit is already lined up.