@oleg-nenashev 2019-12-30
Should we update the code to run publishing only once?