@jglick 2020-01-02
> Should we update the code to run publishing only once?

Not sure what you mean by that.